ipl-attendance/requirement.txt
cmake==3.18.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pytz==2020.1